Generic, nonintrusive, BreadthFirstSearch
Monodirectional breadth first search. uses stl, auto, and functional, implies, but does not require lambdas.
Usage: 
std::stack<keyType> BreadthFirstSearch<keyType = int, invalidKeyValue = -1>::FindPath(startIndex, endIndex, getChildren);
Where: 
keyType is the type of the key used to denote a specific node
invalidKeyValue is the keyType value used to denote a key which has no associated node
startIndex is a keyType value used to denote the start position of the search
endIndex is a keyType value used to denote the end position of the search
getChildren is a function which, given an keyType value, returns a vector of keyType values representing valid transition nodes from the node denoted by the input keyType value. That is to say a functional representing the signature std::vector<keyType> getChildren(keyType index);

The template has, as a default, integer keys to denote nodes, and uses -1 as the invalid node value.